Implementações
Freq.

+++ application/modules/cursos/views/cursos_geral.php
@@ -8,19 +8,26 @@
     <?php 
     echo br(2);
     echo "<table cellpadding='5'>"; 
-       
-        foreach ($cursos as $curso) {
-            echo "<tr>";
-                echo "<td width='550'>".$curso->nome."</td>";
-                
-                $img = img('asset/img/edit.png');
-                echo "<td>".anchor('cursos/editar/'.$curso->ID_CURSO,$img)."</td>";
-                
-                $img = img('asset/img/del.png');
-                echo "<td>".anchor('cursos/excluir/'.$curso->ID_CURSO,$img,
-                        "onclick=\"return confirm('Confirma a exlusão do curso: $curso->nome ? ')\"")."</td>";
-                
-            echo "</tr>";  
+        if ($cursos){
+            foreach ($cursos as $curso) {
+                echo "<tr>";
+                    echo "<td width='550'>".$curso->nome."</td>";
+
+                    $img = img('asset/img/edit.png');
+                    echo "<td>".anchor('cursos/editar/'.$curso->ID_CURSO,$img)."</td>";
+
+                    $img = img('asset/img/del.png');
+                    echo "<td>".anchor('cursos/excluir/'.$curso->ID_CURSO,$img,
+                            "onclick=\"return confirm('Confirma a exlusão do curso: $curso->nome ? ')\"")."</td>";
+
+                echo "</tr>";  
+            }    
+        }
+        else{
+             echo br(2);
+             echo "<tr>";
+                echo "<td width='550'><center>Nenhum curso cadastrado</center></td>";
+             echo "</tr>";
         }
     echo "</table>";
