+++ application/modules/horarios/views/horarios_geral.php
@@ -8,20 +8,27 @@
     <?php 
     echo br(2);
     echo "<table cellpadding='5' >"; 
-       
-        foreach ($horarios as $horario) {
-            echo "<tr>";
-                echo "<td width='300'>".$horario->dia."</td>";
-                echo "<td width='240'>".$horario->hora."</td>";
-                
-                $img = img('asset/img/edit.png');
-                echo "<td>".anchor('horarios/editar/'.$horario->ID_horarios,$img)."</td>";
-                
-                $img = img('asset/img/del.png');
-                echo "<td>".anchor('horarios/excluir/'.$horario->ID_horarios,$img,
-                        "onclick=\"return confirm('Confirma a exlusão do horário: $horario->dia, $horario->hora ? ')\"")."</td>";
-                
-            echo "</tr>";  
+        if ($horarios){
+            foreach ($horarios as $horario) {
+                echo "<tr>";
+                    echo "<td width='300'>".$horario->dia."</td>";
+                    echo "<td width='240'>".$horario->hora."</td>";
+
+                    $img = img('asset/img/edit.png');
+                    echo "<td>".anchor('horarios/editar/'.$horario->ID_horarios,$img)."</td>";
+
+                    $img = img('asset/img/del.png');
+                    echo "<td>".anchor('horarios/excluir/'.$horario->ID_horarios,$img,
+                            "onclick=\"return confirm('Confirma a exlusão do horário: $horario->dia, $horario->hora ? ')\"")."</td>";
+
+                echo "</tr>";  
+            }
+        }
+        else{
+             echo br(2);
+             echo "<tr>";
+                echo "<td width='550'><center>Nenhum horário cadastrado</center></td>";
+             echo "</tr>";
         }
